Generate android_library, android_binary and java_library targets, and make them buildable with Bazel
