+++ src/main/kotlin/com/google/androidstudiopoet/generators/android_modules/AndroidModuleBuildBazelGenerator.kt
@@ -14,16 +14,15 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-package com.google.androidstudiopoet.generators
+package com.google.androidstudiopoet.generators.android_modules
 
-import com.google.androidstudiopoet.models.ModuleBuildSpecificationBlueprint
-import com.google.androidstudiopoet.models.ModuleDependency
+import com.google.androidstudiopoet.models.*
 import com.google.androidstudiopoet.writers.FileWriter
 
-class ModuleBazelBuildGenerator(private val fileWriter: FileWriter) {
-    fun generate(blueprint: ModuleBuildSpecificationBlueprint) {
+class AndroidModuleBuildBazelGenerator(val fileWriter: FileWriter) {
+    fun generate(bazelBlueprint: AndroidBuildBazelBlueprint) {
 
-        val deps: Set<String> = blueprint.dependencies.mapNotNull {
+        val deps: Set<String> = bazelBlueprint.dependencies.mapNotNull {
             when (it) {
                 is ModuleDependency -> "\"//${it.name}\""
                 else -> ""
@@ -33,15 +32,19 @@ class ModuleBazelBuildGenerator(private val fileWriter: FileWriter) {
     deps = [
         ${deps.joinToString(separator = ",\n        ") { it }}
     ],"""
-        val ruleClass = "java_library"
-        val targetName = blueprint.moduleName
+        val ruleClass = if (bazelBlueprint.isApplication) "android_binary" else "android_library"
+        val targetName = bazelBlueprint.packageName.split(".")[1]
         val ruleDefinition = """$ruleClass(
     name = "$targetName",
     srcs = glob(["src/main/java/**/*.java"]),
+    resource_files = glob(["src/main/res/**/*"]),
+    manifest = "src/main/AndroidManifest.xml",
+    custom_package = "${bazelBlueprint.packageName}",
     visibility = ["//visibility:public"], ${if (deps.isNotEmpty()) depsString else ""}
 )
         """
 
-        fileWriter.writeToFile(ruleDefinition, blueprint.path)
+        fileWriter.writeToFile(ruleDefinition, bazelBlueprint.path)
     }
 }
+
