+++ src/main/kotlin/com/google/androidstudiopoet/models/ModuleBuildSpecificationBlueprint.kt
@@ -2,6 +2,7 @@ package com.google.androidstudiopoet.models
 
 interface ModuleBuildSpecificationBlueprint {
 
+    val moduleName: String
     val plugins: Set<String>
     val path: String
     val dependencies: Set<Dependency>
